Move RowCursor back
